+++ layers/GatoGraphQLForWP/plugins/gato-graphql/CHANGELOG.md
@@ -24,6 +24,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - The Settings page has been re-designed, featuring 2-level organization and tabs displayed vertically
 - Reset settings, and choose to use restrictive or non-restrictive default settings
 - Configuration blocks in the the Schema Configuration CPT editor can be removed (and added again)
+- GraphiQL clients now accept sending headers
 - Display warnings in the GraphQL response
 - Implementation of standard custom scalar types
 - Added internal endpoints to feed data to (Gutenberg) blocks